+++ src/script/bindings/mod.rs
@@ -4,6 +4,7 @@ pub mod dict;
 pub mod dns;
 pub mod helpers;
 pub mod http;
+pub mod ldap;
 pub mod ntp;
 pub mod ping;
 pub mod redis;